Fetch token pair historical data via tRPC (#2700)
* add token pair query and tRPC procedure

* add min and max to procedure

* fix sidecar pools query

* add historical price to add CL config

* remove code and fix currencies

* move historical & depth chart to tRPC

* fix build: imports

* delete old store

* fix build

* calculate TVL from pool reserves
