+++ packages/stores/src/queries-external/index.ts
@@ -10,7 +10,6 @@ export * from "./pools";
 export * from "./store";
 export * from "./token-data";
 export * from "./token-historical-chart";
-export * from "./token-pair-historical-chart";
 
 export const IMPERATOR_TIMESERIES_DEFAULT_BASEURL =
   "https://api-osmosis.imperator.co";
