--- packages/stores/src/queries-external/token-pair-historical-chart/types.ts
@@ -1,12 +1,0 @@
-export interface TokenPairHistoricalPrice {
-  close: number;
-  high: number;
-  low: number;
-  open: number;
-  /**
-   * Unix timestamp in seconds
-   */
-  time: number;
-}
-
-export type PriceRange = "1h" | "1d" | "7d" | "1mo" | "1y" | "all";
