+++ packages/web/components/localization/translate-error.ts
@@ -15,7 +15,6 @@ import {
   DepositNoBalanceError,
   HighSwapFeeError,
   InsufficientBalanceError,
-  InvalidRangeError,
   InvalidScalingFactorControllerAddress,
   InvalidSlippageError,
   InvalidSwapFeeError,
@@ -31,6 +30,7 @@ import {
   ScalingFactorTooLowError,
 } from "@osmosis-labs/stores";
 
+import { InvalidRangeError } from "~/hooks";
 import { t } from "~/hooks";
 
 /** Returns localization key given a custom Error subclass, typically from UI configs. */
