+++ packages/web/components/pool-detail/concentrated.tsx
@@ -76,6 +76,7 @@ export const ConcentratedLiquidityPool: FunctionComponent<{ poolId: string }> =
 
     const {
       pool,
+      currentPrice,
       xRange,
       yRange,
       lastChartData,
@@ -90,11 +91,7 @@ export const ConcentratedLiquidityPool: FunctionComponent<{ poolId: string }> =
         poolId,
         priceStore
       ).volume24h;
-    const poolLiquidity = pool?.computeTotalValueLocked(priceStore);
-
-    const currentPrice = pool?.concentratedLiquidityPoolInfo
-      ? pool.concentratedLiquidityPoolInfo.currentPrice
-      : undefined;
+    const poolLiquidity = pool?.totalFiatValueLocked;
 
     const userPositionsInPool = osmosisQueries.queryAccountsPositions
       .get(account?.address ?? "")
@@ -128,9 +125,8 @@ export const ConcentratedLiquidityPool: FunctionComponent<{ poolId: string }> =
       const liquidityUSD = poolLiquidity
         ? Number(poolLiquidity?.toDec().toString())
         : undefined;
-      const poolName = pool?.poolAssets
-        ?.map((poolAsset) => poolAsset.amount.denom)
-        .join(" / ");
+      const poolAssetDenoms = pool?.reserveCoins.map((coin) => coin.denom);
+      const poolName = poolAssetDenoms?.join(" / ");
       const positionCount = userPositionsInPool.length;
 
       logEvent([
@@ -189,21 +185,21 @@ export const ConcentratedLiquidityPool: FunctionComponent<{ poolId: string }> =
                 <div className="flex flex-wrap items-center gap-2">
                   <PoolAssetsIcon
                     className="!w-[78px]"
-                    assets={pool?.poolAssets.map((poolAsset) => ({
-                      coinImageUrl: poolAsset.amount.currency.coinImageUrl,
-                      coinDenom: poolAsset.amount.currency.coinDenom,
+                    assets={pool?.reserveCoins.map((coin) => ({
+                      coinImageUrl: coin.currency.coinImageUrl,
+                      coinDenom: coin.currency.coinDenom,
                     }))}
                   />
                   <div className="flex flex-wrap gap-x-2">
                     <PoolAssetsName
                       size="md"
                       className="text-h5 font-h5"
-                      assetDenoms={pool?.poolAssets.map(
-                        (asset) => asset.amount.currency.coinDenom
+                      assetDenoms={pool?.reserveCoins.map(
+                        (asset) => asset.currency.coinDenom
                       )}
                     />
                     <span className="hidden py-1 text-subtitle1 text-osmoverse-100 lg:inline-block">
-                      {pool?.swapFee ? pool.swapFee.toString() : "0%"}{" "}
+                      {pool?.spreadFactor ? pool.spreadFactor.toString() : "0%"}{" "}
                       {t("clPositions.spreadFactor")}
                     </span>
                   </div>
@@ -241,14 +237,16 @@ export const ConcentratedLiquidityPool: FunctionComponent<{ poolId: string }> =
                 <div className="lg:hidden">
                   <PoolDataGroup
                     label={t("clPositions.spreadFactor")}
-                    value={pool?.swapFee ? pool.swapFee.toString() : "0%"}
+                    value={
+                      pool?.spreadFactor ? pool.spreadFactor.toString() : "0%"
+                    }
                   />
                 </div>
               </div>
             </div>
             <div className="flex h-[340px] flex-row">
               <div className="flex-shrink-1 flex w-0 flex-1 flex-col gap-[20px] py-7 sm:py-3">
-                {chartConfig.queryTokenPairPrice.isFetching ? (
+                {chartConfig.isHistoricalDataLoading ? (
                   <Spinner className="m-auto" />
                 ) : !chartConfig.historicalChartUnavailable ? (
                   <>
