+++ packages/web/hooks/ui-config/use-historical-and-depth-data.ts
@@ -2,12 +2,13 @@ import { useEffect, useState } from "react";
 
 import { useStore } from "~/stores";
 import { ObservableHistoricalAndLiquidityData } from "~/stores/derived-data/concentrated-liquidity";
+import { api } from "~/utils/trpc";
 
 export function useHistoricalAndLiquidityData(
   osmosisChainId: string,
   poolId: string
 ): ObservableHistoricalAndLiquidityData {
-  const { queriesExternalStore, queriesStore, chainStore } = useStore();
+  const { queriesStore, chainStore } = useStore();
 
   const osmosisQueries = queriesStore.get(osmosisChainId).osmosis!;
 
@@ -19,8 +20,7 @@ export function useHistoricalAndLiquidityData(
         poolId,
         queriesStore,
         osmosisQueries.queryLiquiditiesPerTickRange.getForPoolId(poolId),
-        osmosisQueries.queryCfmmConcentratedPoolLinks,
-        queriesExternalStore.queryTokenPairHistoricalChart
+        osmosisQueries.queryCfmmConcentratedPoolLinks
       )
   );
   // react dev tools will unmount the component so only dispose if
@@ -33,6 +33,30 @@ export function useHistoricalAndLiquidityData(
     },
     [config]
   );
+
+  const { data: pool } = api.edge.pools.getPool.useQuery({
+    poolId,
+  });
+  if (pool) config.setPool(pool);
+
+  const {
+    data: historicalPriceData,
+    isLoading,
+    isError,
+  } = api.edge.assets.getAssetPairHistoricalPrice.useQuery(
+    {
+      poolId,
+      baseCoinMinimalDenom:
+        pool?.reserveCoins[0].currency.coinMinimalDenom ?? "",
+      quoteCoinMinimalDenom:
+        pool?.reserveCoins[1].currency.coinMinimalDenom ?? "",
+      timeDuration: config.historicalRange,
+    },
+    { enabled: Boolean(pool) }
+  );
+  if (historicalPriceData) config.setHistoricalData(historicalPriceData.prices);
+  config.setIsHistoricalDataLoading(isLoading);
+  config.setHistoricalDataError(isError);
 
   return config;
 }
