+++ packages/web/modals/increase-concentrated-liquidity.tsx
@@ -1,8 +1,5 @@
 import { Dec } from "@keplr-wallet/unit";
-import {
-  ObservableAddConcentratedLiquidityConfig,
-  ObservableQueryLiquidityPositionById,
-} from "@osmosis-labs/stores";
+import { ObservableQueryLiquidityPositionById } from "@osmosis-labs/stores";
 import { observer } from "mobx-react-lite";
 import dynamic from "next/dynamic";
 import Image from "next/image";
@@ -13,10 +10,11 @@ import { MyPositionStatus } from "~/components/cards/my-position/status";
 import { PriceChartHeader } from "~/components/chart/token-pair-historical";
 import { DepositAmountGroup } from "~/components/cl-deposit-input-group";
 import { tError } from "~/components/localization";
-import { useTranslation } from "~/hooks";
 import {
+  ObservableAddConcentratedLiquidityConfig,
   useAddConcentratedLiquidityConfig,
   useConnectWalletModalRedirect,
+  useTranslation,
 } from "~/hooks";
 import { useHistoricalAndLiquidityData } from "~/hooks/ui-config/use-historical-and-depth-data";
 import { ModalBase, ModalBaseProps } from "~/modals/base";
