+++ packages/web/server/queries/complex/pools/providers/sidecar.ts
@@ -1,11 +1,16 @@
-import { CoinPretty, Dec, PricePretty } from "@keplr-wallet/unit";
+import { CoinPretty, PricePretty, RatePretty } from "@keplr-wallet/unit";
 import cachified, { CacheEntry } from "cachified";
 import { LRUCache } from "lru-cache";
 
 import { PoolRawResponse } from "~/server/queries/osmosis";
-import { queryPools } from "~/server/queries/sidecar";
+import {
+  queryPools,
+  UnderlyingConcentratedPool,
+  UnderlyingStablePool,
+  UnderlyingWeightedPool,
+} from "~/server/queries/sidecar";
 
-import { calcAssetValue, getAsset } from "../../assets";
+import { calcSumAssetsValue, getAsset } from "../../assets";
 import { DEFAULT_VS_CURRENCY } from "../../assets/config";
 import { TransmuterPoolCodeIds } from "../env";
 import { Pool, PoolType } from "../index";
@@ -30,25 +35,24 @@ export function getPoolsFromSidecar(): Promise<Pool[]> {
   });
 }
 
+/** Converts a single SQS pool model response to the standard and more useful Pool type. */
 async function makePoolFromSidecarPool(
   sidecarPool: SidecarPool
 ): Promise<Pool | undefined> {
-  const assetValueDec =
-    (await calcAssetValue({
-      anyDenom: "uosmo",
-      amount: sidecarPool.sqs_model.total_value_locked_uosmo,
-    })) ?? new Dec(0);
   const reserveCoins = await getListedReservesFromSidecarPool(sidecarPool);
 
-  // contains unlisted assets
+  // contains unlisted or invalid assets
   if (reserveCoins.length === 0) return;
 
   return {
     id: getPoolIdFromSidecarPool(sidecarPool.underlying_pool),
     type: getPoolTypeFromSidecarPool(sidecarPool.underlying_pool),
     raw: makePoolRawResponseFromUnderlyingPool(sidecarPool.underlying_pool),
+    spreadFactor: new RatePretty(sidecarPool.sqs_model.spread_factor),
     reserveCoins,
-    totalFiatValueLocked: new PricePretty(DEFAULT_VS_CURRENCY, assetValueDec),
+    totalFiatValueLocked: await calcTotalFiatValueLockedFromReserve(
+      reserveCoins
+    ),
   };
 }
 
@@ -76,7 +80,7 @@ export function getPoolTypeFromSidecarPool(
   throw new Error("Unknown pool type: " + JSON.stringify(underlying_pool));
 }
 
-async function getListedReservesFromSidecarPool(
+export async function getListedReservesFromSidecarPool(
   sidecarPool: SidecarPool
 ): Promise<CoinPretty[]> {
   const listedBalances = (
@@ -97,10 +101,49 @@ async function getListedReservesFromSidecarPool(
     )
   ).filter(Boolean) as CoinPretty[];
 
-  if (listedBalances.length !== sidecarPool.sqs_model.balances.length) {
+  // sort denoms per pool type, as SQS returns them in arbitrary order
+  const type = getPoolTypeFromSidecarPool(sidecarPool.underlying_pool);
+  try {
+    if (type === "weighted") {
+      const raw = sidecarPool.underlying_pool as UnderlyingWeightedPool;
+      return raw.pool_assets.map((asset) => {
+        const coin = listedBalances.find(
+          ({ currency: { coinMinimalDenom } }) =>
+            coinMinimalDenom === asset.token.denom
+        );
+        if (!coin)
+          throw new Error("Missing listed balance for " + asset.token.denom);
+        return coin;
+      });
+    } else if (type === "stable") {
+      const raw = sidecarPool.underlying_pool as UnderlyingStablePool;
+      return raw.pool_liquidity.map(({ denom }) => {
+        const coin = listedBalances.find(
+          ({ currency: { coinMinimalDenom } }) => coinMinimalDenom === denom
+        );
+        if (!coin) throw new Error("Missing listed balance for " + denom);
+        return coin;
+      });
+    } else if (type === "concentrated") {
+      const raw = sidecarPool.underlying_pool as UnderlyingConcentratedPool;
+      const token0 = listedBalances.find(
+        ({ currency: { coinMinimalDenom } }) => coinMinimalDenom === raw.token0
+      );
+      const token1 = listedBalances.find(
+        ({ currency: { coinMinimalDenom } }) => coinMinimalDenom === raw.token1
+      );
+
+      if (!token0 || !token1)
+        throw new Error("Missing listed balance for CL pool: " + raw.id);
+
+      return [token0, token1];
+    }
+  } catch (e) {
+    // not listed
     return [];
   }
 
+  // all other cases, sorting is unknown/doesn't matter
   return listedBalances;
 }
 
@@ -108,16 +151,47 @@ async function getListedReservesFromSidecarPool(
 function makePoolRawResponseFromUnderlyingPool(
   underlyingPool: SidecarPool["underlying_pool"]
 ): PoolRawResponse {
+  if ("current_tick_liquidity" in underlyingPool) {
+    return {
+      ...underlyingPool,
+      id: underlyingPool.id?.toString(),
+      current_tick: underlyingPool.current_tick?.toString(),
+      tick_spacing: underlyingPool.tick_spacing?.toString(),
+      exponent_at_price_one: underlyingPool.exponent_at_price_one?.toString(),
+    } as PoolRawResponse;
+  }
+
+  if ("scaling_factors" in underlyingPool) {
+    return {
+      ...underlyingPool,
+      id: underlyingPool.id?.toString(),
+      scaling_factors: underlyingPool.scaling_factors?.map((factor) =>
+        factor.toString()
+      ),
+    } as PoolRawResponse;
+  }
+
   if ("id" in underlyingPool) {
     return {
       ...underlyingPool,
-      id: underlyingPool.id.toString(),
+      id: underlyingPool.id?.toString(),
     } as PoolRawResponse;
   }
 
   return {
     ...underlyingPool,
-    pool_id: underlyingPool.pool_id.toString(),
-    code_id: underlyingPool.code_id.toString(),
+    pool_id: underlyingPool.pool_id?.toString(),
+    code_id: underlyingPool.code_id?.toString(),
   } as PoolRawResponse;
+}
+
+function calcTotalFiatValueLockedFromReserve(reserve: CoinPretty[]) {
+  const assets = reserve.map((coin) => ({
+    anyDenom: coin.currency.coinMinimalDenom,
+    amount: coin.toCoin().amount,
+  }));
+
+  return calcSumAssetsValue({ assets }).then(
+    (value) => new PricePretty(DEFAULT_VS_CURRENCY, value ?? 0)
+  );
 }
