+++ packages/web/server/queries/imperator/index.ts
@@ -3,6 +3,7 @@ export * from "./market-cap";
 export * from "./pools-fees";
 export * from "./token-data";
 export * from "./token-historical-chart";
+export * from "./token-pair-historical-chart";
 export * from "./tokens";
 
 export const IMPERATOR_TIMESERIES_DEFAULT_BASEURL =
