+++ packages/web/stores/derived-data/store.ts
@@ -9,7 +9,6 @@ import {
   ObservableQueryPoolAprs,
   ObservableQueryPoolFeesMetrics,
   ObservableQueryPriceRangeAprs,
-  ObservableQueryTokensPairHistoricalChart,
   OsmosisQueries,
 } from "@osmosis-labs/stores";
 import { DeepReadonly } from "utility-types";
@@ -37,7 +36,6 @@ export class DerivedDataStore extends BaseDerivedDataStore {
       queryActiveGauges: ObservableQueryActiveGauges;
       queryPriceRangeAprs: ObservableQueryPriceRangeAprs;
       queryClPoolAvgAprs: ObservableQueryClPoolAvgAprs;
-      queryTokenPairHistoricalChart: ObservableQueryTokensPairHistoricalChart;
       queryPoolAprs: ObservableQueryPoolAprs;
     },
     protected readonly accountStore: AccountStore<any>,
